Major development locking implementation changes

+++ ch.elexis.core.data/src/ch/elexis/data/Kontakt.java
@@ -398,6 +398,8 @@ public class Kontakt extends PersistentObject {
 	
 	@Override
 	public boolean delete(){
+		if(lockBarrier()) return false;
+		
 		for (Reminder r : getRelatedReminders()) {
 			r.delete();
 		}
